build: Add edl file and update build to do SGX stuff.
The configure script now determines 32 vs 64 bit based on sizeof void
pointer and defines a variable for so automake can select the right
SGX libraries. We also provide a new --enalbe-sim configure option to
allow people to configure the build for the SGX simulator or not.

Not sure how portable this is or whether it will screw up cross
compile stuff (odds are it does) but it's incremental progress.

We're also now processing the EDL file and using the sgx 'edger8r' to
generate the 'trusted' / 'untrusted' code for crossing the domain
boundary.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>