edl: Rename to use '_' instead of '-'.
This is an obnoxious quirk of the SGX SDK / edger8r tool. It generates
code using the file name and it doesn't have a mechanism to replace
characters that are valid in file names but not in C identifiers. This
means that your EDL file can only be named using characters that are valid
C identifiers. The edger8r tool will show a warning when you run afoul of
this but it will still generate the invalid code.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>