Add the application component, call it the 'mgr'.
The design here is simple. For each SGX TCTI created we create another
TCTI outside the application and associate it with an ID assigned to the
TCTI inside the enclave. This is the most simple implementation I could
come up with. The alternative required coming up with a mechanism for
multiplexing access to a single out-going TCTI like a TAB/RM would. This
would get complicated, just like the TAB/RM.

I don't have any solid requirements for multiple independent TCTIs within
the enclave so the memory overhead of maintaining a 1:1 mapping between
TCTIs is expected to be acceptable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>