cmocka: Use mocks that do nothing to get unit tests building again.
This is a really powerful tool, and I'm using it to do effectively nothing.
The problem solved here is that once we start calling the generated functions
from the SGX edger8r we fall down a hole of SGX-isms. What we want to do is to
provide an implementation of the *_ocalls that our theorhetical enclave will
make and hook the enclave library code to call these instead of the generated
SGX ocalls.

The mocks provided here do nothing interesting or even related to testing the
code in the enclave library. They only provide the symbols needed to get the
test code building again. But this is a model that some future test code can
reproduce to test the library code under conditions where the ocall returns
various errors and other weird stuff.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>