Implement a simple state machine with two states to prevent invalid function calls.
This state machine is intended to force the transmit and receive
functions to be called in said order. Calling multiple transmits
before calling receive is currently invalid (thought it may be
possible?). Calling cancel when we're not waiting for a response
makes no sense. Similarly getting a pollHandle only makes sense if
we're waiting for a response. Finally the locality should only be
set if we're not waiting for a resonse (causes ambiguity as to
which locality the outstanding command will be executed with).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>