Fill in the rest of the tests to verify data in the tcti context structure.
This required exposing the symbols for the sgx specific functions.
These aren't something we want users to interact with directly and
so they were given file scope (static). This makes it impossible
to perform these tests though so we had to make a decision to preserve
the encapsulation or to expose these symbols in the name of test
coverage. I'm coming down on the side of test coverage more and
more these days.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>