Remove untested source files from the lcov statistics.
There are two things about this that I don't like:
First I had to use abs_top_srcdir to get the path to files that are
removed from the lcov statistics. For some reason lcov ignored paths
that used just srcdir on account of the relative path (I think).
Second we have no statistics for the utility library that provides
support services to the enclave. This needs to be added but the current
approach to unit testing here prevents this till we have unit tests
for it.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>