build: Fix 'dist' target.
This takes into account the weirdness of the source code that we build
from the EDL. The default behavior of the 'edger8r' tool makes this
situation easier on us since it builds both the .h and .c files. This
allows us to put the .c files in the target specific nodist_*_SOURCES
variable and ignore the missing dependency on the .h file.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>