EDL: Remove some left-over references to the sized_buf_t type.
The fact that the difference between the wrap functions in the unit
tests and the code generated from the EDL don't cause a build-time
error / warning is a friendly reminder: We're performing tricks with the
linker. The compiler doesn't even have an opportunity to check for this.
All we're doing is swapping around symbols when the executable is
linked. In this case the mock function doesn't touch any of the
parameters that were changed so it didn't touch the changed parts of the
function's stack.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>