Remove all inappropriate uses of the 'tss2_' prefix.
The rule I've come to follow over time is to only use the 'tss2_' prefix
/ namespace for things that require it. This is effectively things that
implement a spec from the TSS2 working group from the TCG. The only use
that should remain is for the public symbols from the SGX TCTI header.
This is the 'Tss2_Tcti_Sgx_Init' function and the header where it's
defined.

Nothing else in this repo qualifies under this rule. This includes the
public header for the manager library. We now install it directly into
$(includedir) instead of the 'tss2' subdirectory.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>