mgr: Add finalize function.
This function tears down everything in the static mgr reference. Seems
logical to allow people to deallocate this memory. This is also useful
in the unit tests which have been updated to use this mechanism in a
teardown function that's invoked after each test. This breaks the
inter-test state issues.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>