test: Add test for 'open' failure path.
This requires mocking the 'open' function call which must be done with
care. The tools we use to collect code coverage metrics for our tests
write to files and so they need to be able to call 'open' w/o us getting
in the way. Our solution to this is to only mock calls to open the file
/dev/urandom which the mgr module uses to get random ID values.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>