mgr: Don't abort on failed calls to 'read' in init_ocall.
This could happen for a number of reasons. We should handle this
a bit  more gracefully. Additionally the hash table mutex doesn't need
to be locked so early and so we move it to immediately before the use of
the hash table. This saves us from having to unlock it on the error
path.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>