test: Cover possible failure of 'read' function in init_ocall.
This requires that the call to 'read' be mocked to fail. We return
a known fd that we generate to be unlikely to collide with any fds
that may be created in the application. This value was generated
from my local /dev/urandom using the following pipeline:

cat /dev/urandom | tr -dc 'a-f0-9' | fold -w 8 | head -n 1

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>