build: Replace hard coding of output path for edger8r.
This rule will now place generated source code in the same directory as
the .edl file used to generate it. This requires that we use a GNU Make
specific feature which renders the resultant Makefile "not portable". To
do this portably would require a bunch of duplicate rules that we'd have
to generate at configure time. Since SGX only support a few Linux distros
there's no need for us to be more portable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>