example: Add skeleton and build rules for enclave.
The enclave code should exercise the TCTI so we want to send a command
to the TPM. Initially this can / should just be a pre-canned command
buffer sent over the tcti transmit function. Something from the command
needs to be displayed to the caller or checked against a "known good
value". The vendor ID (uint32_t) is probably the easiest thing we can
do.

Initially however we just need to make sure the enclave is functioning
so this commit just hard codes a value. Once we have the enclave loading
code working and the interaction between the application and the enclave
working we can fill in a real call to the TPM. We also add a printf-like
function so the enclave can send strings out for debugging purposes.
This function is limited to strings under 8k in length (using a constant
from SGX SDK tlibc/stdio.h).

Most of the hard work in this commit is getting the enclave built by
automake. This is difficult because the enclave shared object must be
built in a way that libtool isn't desiged to do so we're left to do it
ourselves. Combine this with having to deal with the BUILT_SOURCES
from the edger8r tool and things get a bit hairy if we're not careful.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>