Fix enclave and support library builds.
I was doing a few things wrong here:
1) the enclave support library (linked into an application, not an
enclave) needed to be built as a static archive. All uses of libtool
have been removed from the build now. All of the libraries we build are
either static libraries or enclaves (weird shared objects that libtool
won't build for us anyways. The map file has been removed as it's only
relevant for shared objects.
2) The ecalls in the tcti_sgx.edl file were not used (not sure why I put
them there in the first place).
3) After removing the ecalls from the edl file we can't generate the
edge routines that we were building into the libraries. We shouldn't
have been doing this anyways so that's good.
4) Now that we don't have any of the generated headers our use of the
ocalls in the implementation of libtss2-tcti-sgx produces a pile of
compiler warnings about implicit declarations. We resolve this by adding
the required declarations directly to the tcti-sgx module.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>