example: Use SGX TCTI to implement getcap_manufacturer.
In the enclave this commit replaces the hard coded value retuned as a
placeholder with an invocation of the SGX TCTI transmit & receive
functions to query the TPM for the property. This is the most simple
implementation sending a pre-canned command buffer and extracting a
single value from a known location in the response buffer.

In the application we now initialize the SGX TCTI manager so that we can
handle ocalls properly. This requires a callback function so that the
manager can instantiate a new "downtstream" TCTI connection. This
function is hard coded to use the tabrmd TCTI which we now depend upon
and check for @ configure time.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>