build: Check for sgx tools at configure time.
We use the typical autoconf macros to get the absolute path of these
tools. We're operating on the assumption that since the SGX SDK bin
directory may or may not actually be on the PATH (an assumption that we
may want to revisit). This requires that we manually add our guess at
the path to the SGX SDK bin directory, and the right one for the
platform (32 vs 64 bits).

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>