build: Use data from SGX SDK environment script.
The previous approach to determining the size of the architecture
address space (32 vs 64 bits). This used a trick to determine the size
of a void pointer on the build platform. This falls flat when cross
compiling. This isn't something that happens for SGX stuff much but it's
still bad form.

This patch removes these sort of checks as well as the --with-sgxsdkdir
option. Instead we now require the caller to have sourced the
'environment' script installed as part of the SGX SDK. Once this has
been sourced we can use hints from the environment to determine paths.
We can also now use pkg-config metadata to get compiler flags and
library paths / names.

No pkg-config meta data is provided for the 'trusted' SGX libraries so
we must still guess at and export the directory that contains the SGX
headers. We must do the same thing for the '-L' option to the linker
when we're statically linking the example enclave and the enclave
library. We do a little trick with pkg-config and the '--libs-only-L'
option.

This allows us to remove most of the sgx_sdk.mk Makefile fragment.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>