build: Remove definition of ENCLAVE_LDFLAGS, set flags in configure.
This will cause a configure failure if any of the requested flags aren't
supported by the linker. Makes the makefiles cleaner and more
flexible. Causes failure earlier in the build process if required flags
aren't supported by the linker.

Note that we force the use of the gold linker in the check. We also
remove the -nodefaultlibs since this isn't supported by the gold linker.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>