tcti-sgx-mgr: Convert tcti_stx_session_t struct to C++ class.
The most complex part of this commit are the changes necessary in the
unit tests. Since thewe are now interacting with C++ objects they have
been converted to cpp files. To keep this from interfering with the
cmocka infrastructure we must wrap the '__wrap' functions in an
'extern "C"' block.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>