tcti-sgx-mgr: Replace tcti_sgx_mgr_t structure with a C++ class.
This class is implemented using the singleton design pattern. This
allows us to get rid of the global instance of the tcti_sgx_mgr_t. The
singleton does however make testing this class a bit difficult since we
cannot reset the state of the singleton after its first instantiation.
This required breaking some of the tests into separate binaries.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>