tcti-sgx-mgr: Replace GLib hash table with std::list
This change is relatively straight forward though it did require a
significant amount of work in the test code. The STL list uses
exceptions to communicate errors where GLib & C are limited to return
values. This disrupts the typical flow of a cmocka unit test and so our
test coverage is likely dragged down a bit by these changes. This will
need to be addressed in the future.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>