tcti-sgx-mgr: Remove last uses of GLib stuff.
This removes a few uses of 'gint' and 'gpointer' types from the code and
test cases. Also we remove all of the build mechanism required to get
GLib headers and libraries linked in properly.
NOTE: The libtcti-sgx-mgr.so still depends on glib and the metadata
reflects this (readelf -d) however this is on account of the dependency
on libtss2-tcti-tabrmd.so.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>