build: Remove example application and enclave from default target.
When code coverage is enabled we won't be able to link the TCTI library
into the enclave. This is caused by missing gcov symbols in the encalve.
If we want to automate the collection of coverage metrics we can't build
the example enclave when 'make check' is run.

This patch removes the example code from the default target. It can now
be built using the 'example' target. While refactoring these build rules
it was found that the existing dependency intended to generate the
enclave_u.c file before building application$(EXEEXT) did not work as
intended.

This patch includes a fix for this, but it's less than ideal. The name
of the object file for application$(EXEEXT) is mangled by autotools a
bit. Accessing this name may be a bit precarious since it's probably not
a supported interface.

Signed-off-by: Philip Tricca <Philip Tricca philip.b.tricca@intel.com>